// src/@fake-db/types.ts
export type IconType = "bing" | "google" | "google_alert" | "custom";
export type Post = {
  id: number;
  title: string;
  published: string;
  link: string;
  category_label: string | null;
  score: number;
  score_second: number;
  source: IconType;
  created_at: string;
  updated_at: string;
  key_word: {
    key_word: string;
  };
  comments: [];
};